requirements.txt
streamlit
ultralytics
keras-facenet
opencv-python
numpy
scikit-learn
Pillow